Fixes the redundant graphs & bluetooth
This commit allows the posterior and posteromedial plots to be graphed and saved.

This commit also changes the way that data is sent to the computer. Data is only sent over bluetooth once it has been requested by the python script. It use to just dump the data which caused synchronization issues but now the data is only sent out once requested.